+++ packages/sdks/angular-sdk/projects/demo-app/src/app/app.component.spec.ts
@@ -10,12 +10,14 @@ jest.mock('@descope/web-js-sdk');
 describe('AppComponent', () => {
   let mockedCreateSdk: jest.Mock;
   const onSessionTokenChangeSpy = jest.fn();
+  const onIsAuthenticatedChangeSpy = jest.fn();
   const onUserChangeSpy = jest.fn();
 
   beforeEach(async () => {
     mockedCreateSdk = mocked(createSdk);
     mockedCreateSdk.mockReturnValue({
       onSessionTokenChange: onSessionTokenChangeSpy,
+      onIsAuthenticatedChange: onIsAuthenticatedChangeSpy,
       onUserChange: onUserChangeSpy
     });