+++ packages/sdks/react-sdk/src/hooks/useSession.ts
@@ -2,8 +2,13 @@ import { useEffect, useMemo, useRef } from 'react';
 import useContext from './useContext';
 
 const useSession = () => {
-  const { session, isSessionLoading, fetchSession, isSessionFetched } =
-    useContext();
+  const {
+    session,
+    isSessionLoading,
+    fetchSession,
+    isSessionFetched,
+    isAuthenticated,
+  } = useContext();
 
   // when session should be received, we want the return value of "isSessionLoading" to be true starting from the first call
   // (and not only when receiving an update from the context)
@@ -14,7 +19,7 @@ const useSession = () => {
     isLoading.current = isSessionLoading;
   }, [isSessionLoading]);
 
-  const shouldFetchSession = !session && !isSessionLoading;
+  const shouldFetchSession = !isAuthenticated && !isSessionLoading;
 
   // we want this to happen before returning a value so we are using "useMemo" and not "useEffect"
   useMemo(() => {
@@ -33,7 +38,7 @@ const useSession = () => {
   return {
     isSessionLoading: isLoading.current,
     sessionToken: session,
-    isAuthenticated: !!session,
+    isAuthenticated,
   };
 };