+++ packages/sdks/web-component/src/lib/descope-wc/BaseDescopeWc.ts
@@ -37,7 +37,6 @@ import {
   ProjectConfiguration,
   SdkConfig,
 } from '../types';
-import initTemplate from './initTemplate';
 
 // this is replaced in build time
 declare const BUILD_VERSION: string;
@@ -63,9 +62,11 @@ class BaseDescopeWc extends BaseClass {
       'redirect-url',
       'auto-focus',
       'store-last-authenticated-user',
+      'refresh-cookie-name',
       'keep-last-authenticated-user-after-logout',
       'validate-on-blur',
       'style-id',
+      'nonce',
     ];
   }
 
@@ -135,8 +136,46 @@ class BaseDescopeWc extends BaseClass {
     this.#initShadowDom();
   }
 
+  #loadInitStyle() {
+    const sheet = new CSSStyleSheet();
+    sheet.replaceSync(`
+    :host {
+      all: initial;
+			width: 100%;
+      display: block;
+		}
+
+		#root {
+			height: 100%;
+      display: flex;
+      flex-direction: column;
+		}
+
+    #content-root {
+      transition: opacity 300ms ease-in-out;
+    }
+
+		#root[data-theme] {
+			background-color: transparent;
+		}
+
+		.fade-out {
+			opacity: 0.1;
+		}
+
+    .hidden {
+      display: none;
+    }
+    `);
+    this.shadowRoot.adoptedStyleSheets ??= [];
+    this.shadowRoot.adoptedStyleSheets = [
+      ...this.shadowRoot.adoptedStyleSheets,
+      sheet,
+    ];
+  }
+
   #initShadowDom() {
-    this.shadowRoot.appendChild(initTemplate.content.cloneNode(true));
+    this.#loadInitStyle();
     this.slotElement = document.createElement('slot');
     this.slotElement.classList.add('hidden');
     this.rootElement.appendChild(this.slotElement);
@@ -190,6 +229,10 @@ class BaseDescopeWc extends BaseClass {
     return res === 'true';
   }
 
+  get refreshCookieName() {
+    return this.getAttribute('refresh-cookie-name') || '';
+  }
+
   get keepLastAuthenticatedUserAfterLogout() {
     const res = this.getAttribute('keep-last-authenticated-user-after-logout');
     return res === 'true';
@@ -224,6 +267,7 @@ class BaseDescopeWc extends BaseClass {
       'redirect-url',
       'auto-focus',
       'store-last-authenticated-user',
+      'refresh-cookie-name',
       'keep-last-authenticated-user-after-logout',
       'preview',
       'storage-prefix',
@@ -231,6 +275,7 @@ class BaseDescopeWc extends BaseClass {
       'client',
       'validate-on-blur',
       'style-id',
+      'nonce',
     ];
 
     BaseDescopeWc.observedAttributes.forEach((attr: string) => {
@@ -253,6 +298,7 @@ class BaseDescopeWc extends BaseClass {
       storeLastAuthenticatedUser: this.storeLastAuthenticatedUser,
       keepLastAuthenticatedUserAfterLogout:
         this.keepLastAuthenticatedUserAfterLogout,
+      refreshCookieName: this.refreshCookieName,
       ...BaseDescopeWc.sdkConfigOverrides,
       projectId,
       baseUrl,
@@ -471,7 +517,16 @@ class BaseDescopeWc extends BaseClass {
 
   static descopeUI: any;
 
+  #handleNonce() {
+    if (this.getAttribute('nonce')) {
+      // the key name "DESCOPE_NONCE" is in use also by the web-components-ui
+      // it's used to set Vaadins style tags nonce
+      (window as any).DESCOPE_NONCE = this.getAttribute('nonce');
+    }
+  }
+
   async init() {
+    this.#handleNonce();
     this.flowStatus = 'loading';
     ['ready', 'error', 'success'].forEach((status: FlowStatus) =>
       this.addEventListener(status, () => {
@@ -578,6 +633,10 @@ class BaseDescopeWc extends BaseClass {
       this.#validateAttrs();
 
       const isInitialRun = oldValue === null;
+
+      if (attrName === 'nonce') {
+        this.#handleNonce();
+      }
 
       this.#flowState.update(({ stepId, executionId }) => {
         let newStepId = stepId;